+++ CHANGELOG.md
@@ -8,10 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Please read through the Keep a Changelog (~5min)](https://keepachangelog.com/en/1.0.0/).
 
-## [X.Y.Z] - YYYY-MM-DD
+## [Unreleased] - YYYY-MM-DD
+
+## Added
+
+- `--jinja-variables-file` for `dbx execute`
 
 ## Fixed
 - Support `jobs_api_version` values provided by config in `ApiClient` construction
+- References and wording in the Python template
 
 ----
 > Unreleased changes must be tracked above this line.